Debugger: Retain values in symbol tree nodes and explicitly update them
Previously the SymbolTreeModel::data and SymbolTreeModel::setData
functions would access the VM memory directly, which meant the views
wouldn't get updated properly.

I think the way I'm doing it now is more inline with what Qt's built-in
models are doing e.g. the filesystem model, so I think this is more like
how the API is supposed to be used.

Additionally, this makes it so that visible items automatically get
updated when the value in memory changes, like with the rest of the
debugger (see PR#10860).