Debugger: Don't automatically reset symbol trees while switching tabs
It would previously tear down any nodes the user expanded after
switching tabs, which was annoying.

Also, this will make it so that any empty symbol tree model will be
reset once a second until it's populated. This should make the global
variable symbols appear automatically if the user opens the debugger
before the symbol table is loaded, for example.